requirements.txt
numpy
tqdm
pandas
torchvision
torchsummary
tensorboardX
scikit-learn
opencv-python
albumentations